+++ CESplitView.m
@@ -188,7 +188,7 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 
 // ------------------------------------------------------
-- (void)setAllBackgroundColorWithAlpha:(float)inAlpha
+- (void)setAllBackgroundColorWithAlpha:(CGFloat)inAlpha
 // 全てのテキストビューの背景透明度を設定
 // ------------------------------------------------------
 {
